Clean up files, imports, and delete dead code. Fix issue with history route not fetching correct history information. Rename folders and files.
