+++ src/components/VirtualizedDataDisplay/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Center, Loader, Text } from '@mantine/core';
-import { ListVirtualizerContext } from 'src/utils/contexts/ListVirtualizerContext';
+import { ListVirtualizerContext } from './ListVirtualizer';
 import ListVirtualizer from './ListVirtualizer';
 import { TRPCClientErrorLike } from '@trpc/client';
 import { AppRouter } from 'src/server/routers';
