--- src/hooks/useScrollRestoration/index.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-
-const useScrollRestoration = () => {
-    React.useEffect(() => {
-
-    }, [])
-}
-
-export default useScrollRestoration;
