+++ src/pages/_app.tsx
@@ -7,10 +7,9 @@ import { persistor, store } from '../redux/store';
 import { ColorScheme, ColorSchemeProvider, Global, MantineProvider } from '@mantine/core';
 import AppLayout from '../components/AppLayout';
 import { useLocalStorage } from '@mantine/hooks';
-import { createContext, useState } from 'react';
+import { useState } from 'react';
 import { trpc } from 'src/utils/trpc';
 import { QueryClient, QueryClientProvider } from 'react-query';
-import { get, set, del } from "idb-keyval";
 import { PersistGate } from 'redux-persist/integration/react';
 import { ModalsProvider } from '@mantine/modals';
 import { httpBatchLink } from '@trpc/client/links/httpBatchLink';
