+++ src/pages/index.tsx
@@ -10,11 +10,11 @@ import { shallowEqual, useDispatch, useSelector } from 'react-redux';
 import { downloadedPostsSelector, downloadPosts, PostsState } from 'src/redux/slices';
 import { useModals } from '@mantine/modals';
 import { clearStorePosts } from 'src/redux/store';
-import { SortType, TopTimeSort, RedditSortTypeConversion as SortTypeConversion, TopSorts } from 'src/interfaces/sorts';
+import { SortType, TopTimeSort, RedditSortTypeConversion as SortTypeConversion, TopSorts } from 'src/types/sorts';
 import { sortTypeMap, topSortTypeMap } from 'src/utils/sortOptionsMap';
 import SortSelect from 'src/components/MobileSelect/SortSelect';
 import { useQueryClient } from 'react-query';
-import { Prompt, PromptAndStoriesWithNormalizedReplies } from 'src/interfaces/db';
+import { Prompt, PromptAndStoriesWithNormalizedReplies } from 'src/types/db';
 import AccountDrawer from 'src/components/AccountDrawer';
 import { useUser } from 'src/hooks/useUser';
 import { showDownloadNotification, updateDownloadNotification } from 'src/utils/notifications';
