--- src/pages/later/index.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import HistoryLayout from 'src/layouts/HistoryLayout';
-
-const UserReadLaterPage = () => {
-
-    return (
-        <HistoryLayout historyType='readLater' />
-    )
-}
-
-export default UserReadLaterPage;
