--- src/utils/contexts/ListVirtualizerContext.ts
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-interface ContextValues {
-    remeasure: () => void
-}
-
-export const ListVirtualizerContext = React.createContext<ContextValues | null>(null);
